fix: updated env vars for vercel
